+++ spec/sbmt/kafka_consumer/routing/karafka_v1_consumer_mapper_spec.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require "rails_helper"
+
+describe Sbmt::KafkaConsumer::Routing::KarafkaV1ConsumerMapper do
+  describe ".call" do
+    it "underscores client-id in consumer-group name" do
+      # client_id is "some-name" in kafka_consumer.yml
+      expect(described_class.new.call("consumer-group"))
+        .to eq("some_name_consumer-group")
+    end
+  end
+end